feat: Add CI, CHANGELOG, Ethereum tests, mapping queries, mmap, parallel respond
Closes #2, #3, #4, #5, #6, #7

## GitHub Actions CI (#2)
- Added .github/workflows/ci.yml with build, test, clippy
- Uses rust-cache for faster builds
- Added CI badge to README

## CHANGELOG + v0.1.0 (#3)
- Added CHANGELOG.md with Keep a Changelog format
- Documented initial release features

## Ethereum format integration tests (#4)
- Added tests/ethereum_format.rs with 4 tests
- Tests account (3-word) and storage (1-word) formats
- Tests with realistic Ethereum values

## query_account/query_storage with mappings (#5)
- Added AccountMapping and StorageMapping in ethereum_db
- Binary format: address(20) + index(8) for accounts
- Binary format: address(20) + slot(32) + index(8) for storage
- CLI args --account-mapping and --storage-mapping
- Graceful 'not found' handling

## Memory-mapped database support (#6)
- Added src/pir/mmap.rs with MmapDatabase
- save_shards_binary() for mmap-friendly format
- load_shard_binary() with on-demand loading
- --binary-output flag in setup CLI

## Parallel response optimization (#7)
- Made respond() use rayon parallel by default
- respond_sequential() available for comparison
- Added benches/respond.rs with Criterion benchmarks

All 205 tests passing (195 unit + 6 e2e + 4 ethereum format).

Amp-Thread-ID: https://ampcode.com/threads/T-5ef1ad17-5465-4b6a-8926-029f6f137a54

+++ src/ethereum_db/adapter.rs
@@ -372,7 +372,7 @@ mod tests {
         let mut account_map_file = File::create(dir.path().join("account-mapping.bin")).unwrap();
         let mut storage_map_file = File::create(dir.path().join("storage-mapping.bin")).unwrap();
 
-        let mut current_idx: u32 = 0;
+        let mut current_idx: u64 = 0;
 
         let mut sorted_accounts = accounts.to_vec();
         sorted_accounts.sort_by_key(|(addr, _)| *addr);
@@ -386,7 +386,7 @@ mod tests {
             account_map_file.write_all(addr).unwrap();
             account_map_file.write_all(&current_idx.to_le_bytes()).unwrap();
 
-            current_idx += WORDS_PER_ACCOUNT as u32;
+            current_idx += WORDS_PER_ACCOUNT as u64;
         }
 
         let mut sorted_storage = storage.to_vec();
